# docs/rewards-platform-openapi.yaml
openapi: 3.1.0
info:
  title: Rewards Platform API
  version: 1.0.0-phase1
  description: |
    Phase 1 API for multi-merchant rewards marketplace (Malaysia pilot: KL + Melaka).
    Locked via grill rounds 1–7. See `docs/rewards-platform-prd.md`.
servers:
  - url: https://api.example.com
    description: Production
  - url: https://staging-api.example.com
    description: Staging
  - url: http://localhost:3001
    description: Local
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /v1/redemptions/validate:
    post:
      tags: [Redemptions]
      security:
        - merchantApiKey: []
      summary: Preview redemption (writes audit log; no state change)
      operationId: validateRedemption
      parameters:
        - $ref: '#/components/parameters/terminalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedemptionTokenInput'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedemptionPreview'
        '404':
          description: Invalid token or backup code
        '409':
          description: Already redeemed
        '422':
          description: Expired or wrong merchant
  /v1/redemptions/confirm:
    post:
      tags: [Redemptions]
      security:
        - merchantApiKey: []
      summary: Confirm redemption (idempotent)
      operationId: confirmRedemption
      parameters:
        - $ref: '#/components/parameters/terminalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/RedemptionTokenInput'
                - type: object
                  required: [idempotencyKey]
                  properties:
                    idempotencyKey:
                      type: string
                      format: uuid
                      description: Must be stable on retry; mismatch with same token → 409
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedemptionConfirmed'
        '409':
          description: Already redeemed or token/idempotencyKey mismatch
        '422':
          description: Expired or wrong merchant
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    merchantApiKey:
      type: http
      scheme: bearer
      description: Merchant org API key
  parameters:
...
    terminalId:
      name: X-Terminal-Id
      in: header
      required: true
      schema:
        type: string
      description: POS terminal identifier for audit
...
  responses:
...
    RateLimited:
      description: Rate limit exceeded
  schemas:
...
    RewardType:
      type: string
      enum: [discount, free_item]
...
    RedemptionTokenInput:
      type: object
      properties:
        token:
          type: string
          description: Opaque QR token
        backupCode:
          type: string
          minLength: 8
          maxLength: 8
      description: Provide token OR backupCode
    RedemptionPreview:
      type: object
      properties:
        claimId:
          type: string
        rewardTitle:
          type: string
        rewardType:
          $ref: '#/components/schemas/RewardType'
        claimExpiresAt:
          type: string
          format: date-time
        valid:
          type: boolean
    RedemptionConfirmed:
      type: object
      properties:
        redemptionId:
          type: string
        claimId:
          type: string
        redeemedAt:
          type: string
          format: date-time
        idempotencyKey:
          type: string
...